Moved the implementation in Fgs2 into Fgs and got rid of Fgs2. Also, fixed a rarely occurring bug in Gfci. Also, quieted the output for the Maven test run.
