+++ tetrad-gui/src/main/java/edu/cmu/tetradapp/model/ImagesRunner.java
@@ -48,7 +48,7 @@ public class ImagesRunner extends AbstractAlgorithmRunner implements IFgsRunner,
     private transient List<PropertyChangeListener> listeners;
     private List<ScoredGraph> topGraphs;
     private int index;
-    private transient Fgs2 fgs;
+    private transient Fgs fgs;
     private Graph graph;
     private FgsRunner.Type type;
 
@@ -309,7 +309,7 @@ public class ImagesRunner extends AbstractAlgorithmRunner implements IFgsRunner,
 
         if (model instanceof Graph) {
             GraphScore gesScore = new GraphScore((Graph) model);
-            fgs = new Fgs2(gesScore);
+            fgs = new Fgs(gesScore);
             fgs.setKnowledge(getParams().getKnowledge());
             fgs.setNumPatternsToStore(params.getIndTestParams().getNumPatternsToSave());
             fgs.setVerbose(true);
@@ -319,7 +319,7 @@ public class ImagesRunner extends AbstractAlgorithmRunner implements IFgsRunner,
             if (dataSet.isContinuous()) {
                 SemBicScore gesScore = new SemBicScore(new CovarianceMatrixOnTheFly((DataSet) model));
                 gesScore.setPenaltyDiscount(params.getComplexityPenalty());
-                fgs = new Fgs2(gesScore);
+                fgs = new Fgs(gesScore);
                 fgs.setKnowledge(getParams().getKnowledge());
                 fgs.setNumPatternsToStore(params.getIndTestParams().getNumPatternsToSave());
                 fgs.setFaithfulnessAssumed(((FgsIndTestParams) params.getIndTestParams()).isFaithfulnessAssumed());
@@ -330,7 +330,7 @@ public class ImagesRunner extends AbstractAlgorithmRunner implements IFgsRunner,
                 BDeuScore score = new BDeuScore(dataSet);
                 score.setSamplePrior(samplePrior);
                 score.setStructurePrior(structurePrior);
-                fgs = new Fgs2(score);
+                fgs = new Fgs(score);
                 fgs.setVerbose(true);
                 fgs.setKnowledge(getParams().getKnowledge());
                 fgs.setNumPatternsToStore(params.getIndTestParams().getNumPatternsToSave());
@@ -341,7 +341,7 @@ public class ImagesRunner extends AbstractAlgorithmRunner implements IFgsRunner,
         } else if (model instanceof ICovarianceMatrix) {
             SemBicScore gesScore = new SemBicScore((ICovarianceMatrix) model);
             gesScore.setPenaltyDiscount(params.getComplexityPenalty());
-            fgs = new Fgs2(gesScore);
+            fgs = new Fgs(gesScore);
             fgs.setKnowledge(getParams().getKnowledge());
             fgs.setNumPatternsToStore(params.getIndTestParams().getNumPatternsToSave());
             fgs.setFaithfulnessAssumed(((FgsIndTestParams) params.getIndTestParams()).isFaithfulnessAssumed());
@@ -370,11 +370,11 @@ public class ImagesRunner extends AbstractAlgorithmRunner implements IFgsRunner,
                 if (indTestParams.isFirstNontriangular()) {
                     SemBicScoreImages fgsScore = new SemBicScoreImages(list);
                     fgsScore.setPenaltyDiscount(penalty);
-                    fgs = new Fgs2(fgsScore);
+                    fgs = new Fgs(fgsScore);
                 } else {
                     SemBicScoreImages fgsScore = new SemBicScoreImages(list);
                     fgsScore.setPenaltyDiscount(penalty);
-                    fgs = new Fgs2(fgsScore);
+                    fgs = new Fgs(fgsScore);
                 }
             } else if (allDiscrete(list)) {
                 double structurePrior = ((FgsParams) getParams()).getStructurePrior();
@@ -385,9 +385,9 @@ public class ImagesRunner extends AbstractAlgorithmRunner implements IFgsRunner,
                 fgsScore.setStructurePrior(structurePrior);
 
                 if (indTestParams.isFirstNontriangular()) {
-                    fgs = new Fgs2(fgsScore);
+                    fgs = new Fgs(fgsScore);
                 } else {
-                    fgs = new Fgs2(fgsScore);
+                    fgs = new Fgs(fgsScore);
                 }
             } else {
                 throw new IllegalArgumentException("Data must be either all discrete or all continuous.");
