+++ tetrad-lib/src/main/java/edu/cmu/tetrad/performance/Comparison.java
@@ -208,7 +208,7 @@ public class Comparison {
             result.setCorrectResult(SearchGraphUtils.patternForDag(trueDag));
         } else if (params.getAlgorithm() == ComparisonParameters.Algorithm.FGS2) {
             if (score == null) throw new IllegalArgumentException("Score not set.");
-            Fgs2 search = new Fgs2(score);
+            Fgs search = new Fgs(score);
             search.setFaithfulnessAssumed(params.isOneEdgeFaithfulnessAssumed());
             result.setResultGraph(search.search());
             result.setCorrectResult(SearchGraphUtils.patternForDag(trueDag));
