+++ tetrad-lib/src/test/java/edu/cmu/tetrad/test/TestFgs.java
@@ -85,7 +85,7 @@ public class TestFgs {
         SemBicScore score = new SemBicScore(cov);
         score.setPenaltyDiscount(penaltyDiscount);
 
-        Fgs2 fgs = new Fgs2(score);
+        Fgs fgs = new Fgs(score);
         fgs.setVerbose(false);
         fgs.setNumPatternsToStore(0);
         fgs.setOut(out);
@@ -152,7 +152,7 @@ public class TestFgs {
         score.setSamplePrior(samplePrior);
         score.setStructurePrior(structurePrior);
 
-        Fgs2 ges = new Fgs2(score);
+        Fgs ges = new Fgs(score);
         ges.setVerbose(false);
         ges.setNumPatternsToStore(0);
         ges.setFaithfulnessAssumed(false);
@@ -187,7 +187,7 @@ public class TestFgs {
     @Test
     public void testExplore3() {
         Graph graph = GraphConverter.convert("A-->B,A-->C,B-->D,C-->D");
-        Fgs2 fgs = new Fgs2(new GraphScore(graph));
+        Fgs fgs = new Fgs(new GraphScore(graph));
         Graph pattern = fgs.search();
         assertEquals(SearchGraphUtils.patternForDag(graph), pattern);
     }
@@ -195,7 +195,7 @@ public class TestFgs {
     @Test
     public void testExplore4() {
         Graph graph = GraphConverter.convert("A-->B,A-->C,A-->D,B-->E,C-->E,D-->E");
-        Fgs2 fgs = new Fgs2(new GraphScore(graph));
+        Fgs fgs = new Fgs(new GraphScore(graph));
         Graph pattern = fgs.search();
         assertEquals(SearchGraphUtils.patternForDag(graph), pattern);
     }
@@ -203,7 +203,7 @@ public class TestFgs {
     @Test
     public void testExplore5() {
         Graph graph = GraphConverter.convert("A-->B,A-->C,A-->D,A->E,B-->F,C-->F,D-->F,E-->F");
-        Fgs2 fgs = new Fgs2(new GraphScore(graph));
+        Fgs fgs = new Fgs(new GraphScore(graph));
         fgs.setFaithfulnessAssumed(false);
         Graph pattern = fgs.search();
         assertEquals(SearchGraphUtils.patternForDag(graph), pattern);
@@ -211,13 +211,13 @@ public class TestFgs {
 
     @Test
     public void testFromGraph() {
-        int numNodes = 20;
+        int numNodes = 10;
         int numIterations = 10;
 
         for (int i = 0; i < numIterations; i++) {
 //            System.out.println("Iteration " + (i + 1));
             Graph dag = GraphUtils.randomDag(numNodes, 0, numNodes, 10, 10, 10, false);
-            Fgs3 fgs = new Fgs3(new GraphScore(dag));
+            Fgs fgs = new Fgs(new GraphScore(dag));
             fgs.setFaithfulnessAssumed(false);
             Graph pattern1 = fgs.search();
             Graph pattern2 = new Pc(new IndTestDSep(dag)).search();
@@ -248,12 +248,12 @@ public class TestFgs {
         g.addDirectedEdge(x4, x3);
 
         Graph pattern1 = new Pc(new IndTestDSep(g)).search();
-        Fgs2 fgs = new Fgs2(new GraphScore(g));
+        Fgs fgs = new Fgs(new GraphScore(g));
         fgs.setFaithfulnessAssumed(false);
         Graph pattern2 = fgs.search();
 
-        System.out.println(pattern1);
-        System.out.println(pattern2);
+//        System.out.println(pattern1);
+//        System.out.println(pattern2);
 
         assertEquals(pattern1, pattern2);
     }
@@ -284,8 +284,8 @@ public class TestFgs {
         fgs.setFaithfulnessAssumed(false);
         Graph pattern2 = fgs.search(x1);
 
-        System.out.println(pattern1);
-        System.out.println(pattern2);
+//        System.out.println(pattern1);
+//        System.out.println(pattern2);
 
         assertEquals(pattern1, pattern2);
     }
@@ -300,7 +300,7 @@ public class TestFgs {
             Graph dag = GraphUtils.randomDag(numNodes, 0, numNodes, 10, 10, 10, false);
             GraphScore fgsScore = new GraphScore(dag);
 
-            Fgs2 fgs = new Fgs2(fgsScore);
+            Fgs fgs = new Fgs(fgsScore);
             Graph pattern1 = fgs.search();
 
             Node x1 = fgsScore.getVariable("X1");
@@ -438,7 +438,7 @@ public class TestFgs {
 
         SemBicScore score = new SemBicScore(dataSet);
         score.setPenaltyDiscount(1);
-        Fgs2 fgs = new Fgs2(score);
+        Fgs fgs = new Fgs(score);
         fgs.setKnowledge(knowledge);
 
         Graph pattern = fgs.search();
@@ -480,7 +480,7 @@ public class TestFgs {
         Graph graph = GraphConverter.convert(inputGraph);
 
         // Set up search.
-        Fgs2 fgs = new Fgs2(new GraphScore(graph));
+        Fgs fgs = new Fgs(new GraphScore(graph));
 
         // Run search
         Graph resultGraph = fgs.search();
@@ -512,7 +512,7 @@ public class TestFgs {
         Graph input = GraphConverter.convert(inputGraph);
 
         // Set up search.
-        Fgs2 fgs = new Fgs2(new GraphScore(input));
+        Fgs fgs = new Fgs(new GraphScore(input));
 
         // Set up search.
         fgs.setKnowledge(knowledge);
